Change hour props from int to decimal
Maybe hour is wrong, should prob chng to period or something

+++ src/SolarTally.Domain/Interfaces/IReadOnlySiteSettings.cs
@@ -4,7 +4,7 @@ namespace SolarTally.Domain.Interfaces
 {
     public interface IReadOnlySiteSettings
     {
-        int NumSolarHours { get; }
+        decimal NumSolarHours { get; }
         TimeInterval PeakSolarInterval { get; }
     }
 }
